refactor(output): extract native JXL strategy into Output.NativeJxlSearch
The native-JXL distance strategy (clamp + single encode / max_bytes self-cap, no
band loop) shared nothing with EncodeSearch's measure-and-search machinery but
lived inside it (~90 lines). Move it to its own module; EncodeSearch.run/3 keeps a
one-line delegating clause so the unified {:ok, binary, meta} execution entry and
the [:encode, :search] span are preserved. Acyclic: EncodeSearch -> NativeJxlSearch
at runtime, NativeJxlSearch -> EncodeSearch only for the meta() type. No behaviour
change; native cases pass through the delegation.
